Split parser into Loader and parser
Fixes #47
