+++ src/test/java/db/data/load/CsvParserTest.java
@@ -3,11 +3,9 @@ package db.data.load;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.io.FileInputStream;
-import java.io.IOException;
 import java.util.ArrayList;
 
 import db.structure.recherches.TableHandler;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -26,7 +24,7 @@ import db.structure.Column;
 import db.structure.Table;
 
 class CsvParserTest {
-	protected Parser parser;
+	protected Loader loader;
 	protected Table table;
 	Utils utilsInstance;
 
@@ -66,24 +64,16 @@ class CsvParserTest {
 		columns.stream().forEach(c -> tableHandler.addColumn(c));
 
 		Table table = tableHandler.createTable();
-		//parser = new CsvParser2(table); // 790 ms
-		parser = new CsvParser(table); // 750 ms
-	}
-
-	@AfterEach
-	void tearDown() throws Exception {
+		loader = new Loader(table, new CsvParser()); // 750 ms
 	}
 
 	@Test
-	void testParseInputStreamInt() throws IOException {
-		Executable exec = new Executable() {
-			@Override
-			public void execute() throws Throwable {
-				FileInputStream is = new FileInputStream("testdata/SMALL_100_000_yellow_tripdata_2015-04.csv");
-				Timer parseTimer = new Timer("Temps pris par le parsing");
-				parser.parse(is, false); // Supprimer les données déjà écrites
-				parseTimer.log();
-			}
+	void testParseInputStreamInt() {
+		Executable exec = () -> {
+			FileInputStream is = new FileInputStream("testdata/SMALL_100_000_yellow_tripdata_2015-04.csv");
+			Timer parseTimer = new Timer("Temps pris par le parsing");
+			loader.parse(is, false); // Supprimer les données déjà écrites
+			parseTimer.log();
 		};
 		assertDoesNotThrow(exec);
